requirements.txt
asgiref==3.3.4
astroid==2.5.3
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.14
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.11.2
idna==2.10
isort==5.8.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==2.2.2
pylint==2.7.4
pylint-django==2.4.3
pylint-plugin-utils==0.6
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
gunicorn
--install-option="--no-deps" django-heroku
dj-database-url
psycopg2-binary
whitenoise